Renamed email dispatcher, moved package to utils
